gateway: split the Chat-Completions frame mapping out of openai.rs
Mechanical move for the module line budget after the forensics additions:
feed_openai_compatible and its [DONE] terminal handling relocate to
dialects/openai/compatible.rs unchanged.
